TpExecutorProps add dtp field for judge if create dtp executor or not
